ParticleSet: Move ParticleSet into separate file
