+++ parcels/__init__.py
@@ -1,5 +1,6 @@
 from parcels.grid import *  # noqa
 from parcels.particle import *  # noqa
+from parcels.particleset import *  # noqa
 from parcels.field import *  # noqa
 from parcels.kernel import *  # noqa
 import parcels.rng as random  # noqa
